+++ amazonorders/selectors.py
@@ -86,22 +86,33 @@ class Selectors:
                                      "div.order"]
     ORDER_HISTORY_COUNT_SELECTOR = ".js-yo-container span.num-orders"
     ORDER_DETAILS_ENTITY_SELECTOR = ["div#orderDetails",
-                                     "div#ordersContainer"]
+                                     "div#ordersContainer",
+                                     "div#odp-main-section"]
     ITEM_ENTITY_SELECTOR = ["[data-component='purchasedItems'] .a-fixed-left-grid",
                             "div:has(> div.yohtmlc-item)",
-                            ".item-box"]
+                            ".item-box",
+                            # WFM in-store line items
+                            "div.a-row.a-spacing-base:has(img.ufpo-itemListWidget-image)"]
     SHIPMENT_ENTITY_SELECTOR = ["[data-component='orderCard'] [data-component='shipments'] .a-box",
                                 "div.shipment",
                                 "div.delivery-box"]
     # Selectors defined here mean we don't have a reliable way to parse all details in an Order, so Items and
     # Shipments will be skipped
     ORDER_SKIP_ITEMS = [
-        # Identifies an Amazon Fresh order
+        # Identifies an Amazon Fresh order (also matched by WFM in-store; distinguished via ORDER_WHOLE_FOODS)
         ".brand-info-box .brand-logo img",
-        # Identifies a Whole Foods Market order
+        # Identifies a Whole Foods Market receipt order
         "a.yohtmlc-order-details-link[href^='/wholefoodsmarket']",
         # Identifies an order from a physical Amazon store
         Selector("div.yohtmlc-shipment-status-primaryText", "Purchased at Amazon")
+    ]
+    # Selectors identifying a Whole Foods Market purchase. Unlike ORDER_SKIP_ITEMS entries, WFM orders
+    # expose a grand_total (and often item_count) on the history page, so those fields are populated.
+    ORDER_WHOLE_FOODS = [
+        "a[href*='/wholefoodsmarket/receipts/order/']",
+        "a[href*='/fopo/order-details']",
+        Selector("div.yohtmlc-shipment-status-primaryText", text_contains="Whole Foods Market"),
+        "img.ufpo-itemListWidget-image"
     ]
     # Selectors defined here mean the Order will not have parsable totals
     ORDER_SKIP_TOTALS = [
@@ -115,19 +126,28 @@ class Selectors:
     # CSS selectors for Item fields
     #####################################
 
-    FIELD_ITEM_IMG_LINK_SELECTOR = "a img"
+    # Last entry matches WFM in-store line items
+    FIELD_ITEM_IMG_LINK_SELECTOR = ["a img",
+                                    "img.ufpo-itemListWidget-image"]
     FIELD_ITEM_QUANTITY_SELECTOR = [".od-item-view-qty",
                                     "span.item-view-qty",
                                     "span.product-image__qty"]
+    # WFM items render "Qty: N" or "Qty: 0.31 lb"; extracted in Item._parse_quantity
+    FIELD_ITEM_WHOLE_FOODS_QUANTITY_SELECTOR = ["span.a-size-small"]
     FIELD_ITEM_TITLE_SELECTOR = ["[data-component='itemTitle']",
-                                 ".yohtmlc-item a", ".yohtmlc-product-title"]
+                                 ".yohtmlc-item a", ".yohtmlc-product-title",
+                                 "div.a-column.a-span10 > a",
+                                 # ASINLESS WFM items render the title in a span rather than a link
+                                 "div.a-column.a-span10 > span"]
     FIELD_ITEM_LINK_SELECTOR = ["[data-component='itemTitle'] a",
                                 ".yohtmlc-item a",
                                 "a:has(> .yohtmlc-product-title)",
-                                ".yohtmlc-product-title a"]
+                                ".yohtmlc-product-title a",
+                                "div.a-column.a-span10 > a"]
     FIELD_ITEM_TAG_ITERATOR_SELECTOR = [".yohtmlc-item div"]
     FIELD_ITEM_PRICE_SELECTOR = ["[data-component='unitPrice'] .a-text-price :not(.a-offscreen)",
-                                 ".yohtmlc-item .a-color-price"]
+                                 ".yohtmlc-item .a-color-price",
+                                 "div.a-section.a-text-right span.a-size-small"]
     FIELD_ITEM_SELLER_SELECTOR = ["[data-component='orderedMerchant']"] + FIELD_ITEM_TAG_ITERATOR_SELECTOR
     FIELD_ITEM_RETURN_SELECTOR = ["[data-component='itemReturnEligibility']"] + FIELD_ITEM_TAG_ITERATOR_SELECTOR
 
@@ -138,7 +158,9 @@ class Selectors:
     FIELD_ORDER_DETAILS_LINK_SELECTOR = ["a.yohtmlc-order-details-link",
                                          # Would like to use this or similar, but not yet sure how consistent it is
                                          # ".order-header__header-link-list-item:first-of-type a"
-                                         ]
+                                         # WFM receipt and FOPO orders link to dedicated pages, not standard endpoint
+                                         "a[href*='/wholefoodsmarket/receipts/order/']",
+                                         "a[href*='/fopo/order-details']"]
     FIELD_ORDER_NUMBER_SELECTOR = ["[data-component='orderId']",
                                    "[data-component='briefOrderInfo'] div.a-column",
                                    ".order-date-invoice-item :is(bdi, span)[dir='ltr']",
@@ -146,7 +168,13 @@ class Selectors:
                                    ":is(bdi, span)[dir='ltr']"]
     FIELD_ORDER_GRAND_TOTAL_SELECTOR = ["div.yohtmlc-order-total span.value",
                                         "div.order-header div.a-column.a-span2",
-                                        "div.order-header div.a-col-left .a-span9"]
+                                        "div.order-header div.a-col-left .a-span9",
+                                        "#wfm-grand-total-amount"]
+    # WFM in-store (FOPO) details page amounts
+    FIELD_ORDER_WHOLE_FOODS_SUBTOTAL_SELECTOR = "#wfm-subtotal-amount"
+    FIELD_ORDER_WHOLE_FOODS_TAX_SELECTOR = "#wfm-tax-total-amount"
+    FIELD_ORDER_WHOLE_FOODS_PAYMENT_METHOD_SELECTOR = "#wfm-0-card-brand"
+    FIELD_ORDER_WHOLE_FOODS_PAYMENT_LAST_4_SELECTOR = "#wfm-0-card-tail"
     FIELD_ORDER_PLACED_DATE_SELECTOR = ["[data-component='orderDate']",
                                         "span.order-date-invoice-item",
                                         "[data-component='briefOrderInfo'] div.a-column",
@@ -162,6 +190,8 @@ class Selectors:
     FIELD_ORDER_ADDRESS_FALLBACK_1_SELECTOR = "div.recipient span.a-declarative"
     FIELD_ORDER_ADDRESS_FALLBACK_2_SELECTOR = "script[id^='shipToData']"
     FIELD_ORDER_GIFT_CARD_INSTANCE_SELECTOR = ".gift-card-instance"
+    FIELD_ORDER_ITEM_COUNT_SELECTOR = ["div.a-fixed-left-grid-col.a-col-right span",
+                                       "span"]
 
     #####################################
     # CSS selectors for Shipment fields